Added coordinator tests and cleanup
