+++ packages/live-share-media/package.json
@@ -9,7 +9,7 @@
     "scripts": {
         "clean": "npx shx rm -rf bin",
         "build": "tsc -p tsconfig.build.json",
-        "test": "ts-mocha --config ./src/test/.mocharc.js src/**/LiveMediaSession.spec.ts --timeout 10000",
+        "test": "ts-mocha --config ./src/test/.mocharc.js src/**/*.spec.ts --timeout 10000",
         "test:debug": "ts-mocha src/**/*.spec.ts --inspect-brk",
         "test:coverage": "nyc --reporter=html --reporter=text --reporter=text-summary npm test"
     },
