+++ packages/live-share-media/src/internals/GroupPlaybackPosition.ts
@@ -169,9 +169,7 @@ export class GroupPlaybackPosition {
         const now = this._liveRuntime.getTimestamp();
         const ignoreBefore = now - this._updateInterval.milliseconds * 2;
         const shouldProject = !this._transportState.track.metadata?.liveStream;
-        this._positions.forEach((value, key) => {
-            const position = this._positions.get(key)!;
-
+        this._positions.forEach((position, _) => {
             // Ignore any old updates
             if (position.timestamp > ignoreBefore) {
                 // Compute projected playback position
